feat: rename ChubaoFS to CubeFS
