+++ BUILD.md
@@ -1,6 +1,6 @@
 # Build CSI driver image
 
-Build the ChubaoFS client and CSI driver using the following commands
+Build the CubeFS client and CSI driver using the following commands
 ```
 make
 ```
