+++ cmd/cubefs.go
@@ -22,16 +22,16 @@ import (
 	"os/signal"
 	"syscall"
 
+	"github.com/cubefs/cubefs-csi/pkg/cubefs"
 	"github.com/golang/glog"
 
-	"github.com/chubaofs/chubaofs-csi/pkg/chubaofs"
 	"github.com/spf13/cobra"
 )
 
 var (
 	endpoint string
 	version  = "1.0.0"
-	conf     chubaofs.Config
+	conf     cubefs.Config
 )
 
 func init() {
@@ -52,7 +52,7 @@ func main() {
 	cmd.Flags().AddGoFlagSet(flag.CommandLine)
 	cmd.PersistentFlags().StringVar(&conf.NodeID, "nodeid", "", "This node's ID")
 	cmd.PersistentFlags().StringVar(&endpoint, "endpoint", "unix:///csi/csi.sock", "CSI endpoint, must be a UNIX socket")
-	cmd.PersistentFlags().StringVar(&conf.DriverName, "drivername", chubaofs.DriverName, "name of the driver (Kubernetes: `provisioner` field in StorageClass must correspond to this value)")
+	cmd.PersistentFlags().StringVar(&conf.DriverName, "drivername", cubefs.DriverName, "name of the driver (Kubernetes: `provisioner` field in StorageClass must correspond to this value)")
 	cmd.PersistentFlags().StringVar(&conf.Version, "version", version, "Driver version")
 	cmd.PersistentFlags().StringVar(&conf.KubeConfig, "kubeconfig", "", "Kubernetes config")
 	cmd.PersistentFlags().BoolVar(&conf.RemountDamaged, "remountdamaged", false,
@@ -68,9 +68,9 @@ func main() {
 }
 
 func handle() {
-	d, err := chubaofs.NewDriver(conf)
+	d, err := cubefs.NewDriver(conf)
 	if err != nil {
-		glog.Errorf("chubaofs.NewDriver error:%v\n", err)
+		glog.Errorf("cubefs.NewDriver error:%v\n", err)
 		os.Exit(1)
 	}
