+++ deploy/storageclass-default.yaml
@@ -3,15 +3,15 @@ apiVersion: storage.k8s.io/v1
 metadata:
   name: cfs-sc
   namespace: default
-provisioner: csi.chubaofs.com
+provisioner: csi.cubefs.com
 allowVolumeExpansion: true
 reclaimPolicy: Delete
 parameters:
   # Resource manager IP address or URL
-  masterAddr: "master-service.chubaofs.svc.cluster.local:17010"
+  masterAddr: "master-service.cubefs.svc.cluster.local:17010"
   # Owner name as authentication
   owner: "csiuser"
-  #  consulAddr: "http://consul-service.chubaofs.svc.cluster.local:8500"
+  #  consulAddr: "http://consul-service.cubefs.svc.cluster.local:8500"
   #  crossZone: "false"
   #  enableToken: "false"
   #  zoneName: ""
